changed shaders to controller objects, Fixed issue where swapbuffers may not be called immediately after scene draw

+++ src/main/kotlin/engine/world/Updatable.kt
@@ -1,5 +1,5 @@
 package engine.world
 
 interface Updatable {
-    fun update(delta: Float)
+    fun update(delta: Double)
 }
